# internal/requests/user_requests.go
package requests

import (
	validation "github.com/go-ozzo/ozzo-validation/v4"
	"github.com/go-ozzo/ozzo-validation/v4/is"
)

const (
	minPathLength = 8
)

type BasicAuth struct {
	Email    string `json:"email" validate:"required" example:"john.doe@example.com"`
	Password string `json:"password" validate:"required" example:"11111111"`
}

func (ba BasicAuth) Validate() error {
	return validation.ValidateStruct(&ba,
		validation.Field(
			&ba.Email,
			validation.Required.Error("Email is required"),
			is.Email.Error("Invalid email address"),
		),
		validation.Field(
			&ba.Password,
			validation.Required.Error("Password is required"),
			validation.Length(minPathLength, 0).
				Error("Password must be at least 8 characters"),
		),
		// validation.Field(
		// 	&ba.Email,
		// 	validation.Required,
		// 	is.Email,
		// ),
		// validation.Field(
		// 	&ba.Password,
		// 	validation.Required,
		// 	validation.Length(minPathLength, 0),
		// ),
	)
}

type LoginRequest struct {
	BasicAuth
}

type RegisterRequest struct {
	BasicAuth
	Name     string `json:"name" validate:"required" example:"John Doe"`
	Username string `json:"username" validate:"required" example:"johnDoe"`
}

func (rr RegisterRequest) Validate() error {
	err := rr.BasicAuth.Validate()
	if err != nil {
		return err
	}

	return validation.ValidateStruct(&rr,
		validation.Field(&rr.Name, validation.Required),
		validation.Field(&rr.Username, validation.Required),
	)
}

type OAuthRequest struct {
	Token string `json:"token" validate:"required"`
}

func (oar OAuthRequest) Validate() error {
	return validation.ValidateStruct(&oar,
		validation.Field(&oar.Token, validation.Required),
	)
}

type RefreshRequest struct {
	Token string `json:"token" validate:"required" example:"refresh_token"`
}